fix: move components to same folder
